64 bit bug fix remove autoconf

+++ black/include/utils.h
@@ -12,6 +12,6 @@
 #define WARN(...) fprintf(stdout, __VA_ARGS__);
 #define ERROR(...) fprintf(stdout, __VA_ARGS__);
 
-void *memory_grow_to_size(void *mem_p, size_t per_block, size_t *mem_cap_ip, size_t mem_size_i);
+void *memory_grow_to_size(void *mem_p, size_t per_block, int *mem_cap_ip, int mem_size_i);
 
 #endif  /* UTILS_H__ */
